UI更新2.1
1. 修复UI中路径无法保存的bug
2. 更改程序的配置逻辑，避免api key泄露
3. 增加帮助信息
4. 新增打包好的exe
