+++ .gitignore
@@ -8,3 +8,4 @@ Data
 APIKey
 Log
 Result
+build
