+++ Main.py
@@ -1,5 +1,5 @@
-import config_loader as config
-from ui import App
+from lib import config_loader as config
+from lib.ui import App
 
 if __name__ == "__main__":
     # 确保配置已加载
