+++ config.json
@@ -2,13 +2,13 @@
     "save_full_log": true,
     "include_requirements_in_prompt": true,
     "include_keywords_in_prompt": false,
-    "DATA_FOLDER": "Data",
-    "APIKEY_FOLDER": "APIKey",
-    "RESULT_FOLDER": "Result",
+    "DATA_FOLDER": "c:/Github/AutoPaperSearch/Data",
+    "APIKEY_FOLDER": "c:/Github/AutoPaperSearch/APIKey",
+    "RESULT_FOLDER": "c:/Github/AutoPaperSearch/Result",
     "LANGUAGE": "zh_CN",
-    "ResearchQuestion": "Explore the caption / subtitle design",
-    "Requirements": "\n1. Should conduct user study.\n2. Should display subtitle / caption.\n",
-    "Keywords": "",
+    "ResearchQuestion": "Your research question.",
+    "Requirements": "1. xxx xxx.\n2. xxx xxx\n",
+    "Keywords": "xxx, xxx, ...",
     "system_prompt": "你是一个学术论文相关性判断专家。你需要判断论文是否与用户的研究问题要找的答案相关。\n\n判断标准（按照数字顺序依次判断）：\n1. 如果你判断摘要的内容与用户的研究问题的相关性很高，则相关。不进入下一步判断。否则进入下一步判断。\n2. 如果用户给出的#要求#为空，则跳过该步骤。如果你判断摘要的内容符合用户的要求，则相关。不进入下一步判断。否则进入下一步判断。\n3. 如果用户给出的#关键词#为空，则跳过该步骤。如果你判断摘要的内容与用户的研究问题的可能答案相关性很不高，但你判断正文极有可能包含关键词，且猜测正文与用户的研究问题相关性很高，则相关。否则不相关。\n\n重要说明：\n- 关键词不区分大小写\n- 注意同义词、缩写词的匹配\n\n你必须以JSON格式输出判断结果。\n\n输出格式示例（相关）：\n{\n    \"relevant\": \"Y\",\n    \"reason\": \"论文探讨了VST技术在混合现实应用中的使用\"\n}\n\n输出格式示例（不相关）：\n{\n    \"relevant\": \"N\", \n    \"reason\": \"论文仅涉及OST技术，未提及VST或视频透视技术\"\n}\n\n注意：relevant字段只能是\"Y\"或\"N\"，reason字段必须是一句简短精炼的理由。\n",
     "deepseek_chat_standard_prices": {
         "input_cache_hit": 0.5,
@@ -31,57 +31,5 @@
         "output": 4.0
     },
     "model_name": "deepseek-chat",
-    "api_base_url": "https://api.deepseek.com",
-    "API_KEYS": [
-        "sk-5aa84a4938a54a6da00f418a676300e2",
-        "sk-e32b210f5ed04ceba2173583df9e39ac",
-        "sk-5e383fecd9d446baacdc9442bcaea760",
-        "sk-2a1c08eafcf148b386c9ca77a078a9f8",
-        "sk-e3fcb402a0e5418cae39f9ef6e9273a4",
-        "sk-b18b0739763c477ab31a7b9314219c74",
-        "sk-b856c499d850495eaa890fa4c6bdcfbe",
-        "sk-79de3c54aa3142ecb93e7b216745a566",
-        "sk-732ebd3573e14018a6baea1aced279b7",
-        "sk-606f0a021dbe4801806a625df135f064",
-        "sk-32cc4695fd3848cf96699db01a10d0fa",
-        "sk-7bb04d02677340ccaeae2702aec94b90",
-        "sk-cb68f6122597477eb4f1e5c49cdf355e",
-        "sk-0d93665038c247d69f991764c2841fed",
-        "sk-891aa8dc94564b829e615f9210077eec",
-        "sk-e4eb25dce4f84e89b6c6d655d44e6ac8",
-        "sk-45d6e226664a4a3eb381c276506a1cd3",
-        "sk-d129b6bfd1094b81a7f6d6905aadf703",
-        "sk-5efa9ea3624946cbba92520c709abffd",
-        "sk-d7fd354945b34edb9742164908473590",
-        "sk-bf4c885fb57c4d97948105d4bd482344",
-        "sk-c423d612129d48a7a999510b6bd8bba2",
-        "sk-26ad5752d5e747439e601228e655af67",
-        "sk-2eb7d2ab39da42949bad6173e3296674",
-        "sk-3e87102459404a0eb892b095db2f778f",
-        "sk-424b3b01e8ef4b109a077f3004e000e1",
-        "sk-403acc635d394180ba33c240fa47871f",
-        "sk-cdf5f54a7aac472aaccad90230e44321",
-        "sk-19f37f8c1c274cc5be4fe40c7610112d",
-        "sk-1cec59e2a35b44f580a301175f2f10cc",
-        "sk-74dd62b8180d4c9fb5d3a50d238f378a",
-        "sk-f7781a0028f14a7cbfd763ec01f1524e",
-        "sk-11ede10df6474dab8cb82fd6a2ba8a68",
-        "sk-e2ad7d062b7a4468ae69bfc5a59cf95a",
-        "sk-a3613d53756d4ca4879198029de81f30",
-        "sk-9494700868bb43feb192ff0975111f68",
-        "sk-463ac6b46ff340e9a374d575d84372f3",
-        "sk-96f66bfda318459f939b4e559ffb5c8b",
-        "sk-8078b6eb1df74537b669a66fc65d2f96",
-        "sk-7946292e8afc4540aff85277738e7b84",
-        "sk-02bd7a2fcc2b4d74ab073c1094320464",
-        "sk-0e9920577e7a478680e9e26f30b3d0df",
-        "sk-68637fb887244949bdb16029f682b818",
-        "sk-594441106ea94e5b86246859c7b07590",
-        "sk-7c272cdec5d64bdea24cd352129da82b",
-        "sk-11fe2b0d64af421699cffd7985af8def",
-        "sk-0f35e922d4194b13bd75f8c32867b892",
-        "sk-5886df6641054e0eb7a4051ec6d3ba83",
-        "sk-172a5af8d29b44ca8d96a6183401012d",
-        "sk-bea64812754a4ee6932a54b779a01ce3"
-    ]
+    "api_base_url": "https://api.deepseek.com"
 }
