+++ language/__init__.py
@@ -1,0 +1,2 @@
+# Language package
+from .language import *
