+++ lib/paper_processor.py
@@ -3,9 +3,9 @@ import time
 import threading
 from datetime import datetime
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from . import Data, SearchPaper
+from . import data, searchpaper
 from . import utils
-import config_loader as config
+from . import config_loader as config
 import json
 
 def process_paper_batch(paper_indices, rq, keywords, requirements, api_key, thread_id, result_file_path, log_file_path, yon_log_file_path, total_papers):
@@ -16,28 +16,28 @@ def process_paper_batch(paper_indices, rq, keywords, requirements, api_key, thre
     batch_tokens = 0
     
     for idx, paper_index in enumerate(paper_indices):
-        if paper_index in Data.paper_data:
+        if paper_index in data.paper_data:
             # 记录单篇论文处理开始时间
             single_start_time = time.time()
             
-            paper = Data.paper_data[paper_index]
+            paper = data.paper_data[paper_index]
             title = paper['title']
             abstract = paper['abstract']
             entry = paper['entry']
             
             try:
-                # 调用SearchPaper中的方法检查相关性，传入对应的API密钥和requirements
-                relevance, tokens, reason, prompt_tokens, completion_tokens, cache_hit, cache_miss = SearchPaper.check_paper_relevance(
+                # 调用searchpaper中的方法检查相关性，传入对应的API密钥和requirements
+                relevance, tokens, reason, prompt_tokens, completion_tokens, cache_hit, cache_miss = searchpaper.check_paper_relevance(
                     rq, keywords, requirements, title, abstract, api_key)
                 
                 # 累加token使用量
                 batch_tokens += tokens
-                with Data.token_lock:
-                    Data.token_used += tokens
-                    Data.prompt_tokens_used += prompt_tokens
-                    Data.completion_tokens_used += completion_tokens
-                    Data.prompt_cache_hit_tokens_used += cache_hit
-                    Data.prompt_cache_miss_tokens_used += cache_miss
+                with data.token_lock:
+                    data.token_used += tokens
+                    data.prompt_tokens_used += prompt_tokens
+                    data.completion_tokens_used += completion_tokens
+                    data.prompt_cache_hit_tokens_used += cache_hit
+                    data.prompt_cache_miss_tokens_used += cache_miss
                 
                 # 计算单篇论文处理时间
                 single_elapsed_time = time.time() - single_start_time
@@ -46,7 +46,7 @@ def process_paper_batch(paper_indices, rq, keywords, requirements, api_key, thre
                 utils.update_progress(single_elapsed_time)
                 
                 # 记录到Y/N日志文件（无论结果是Y还是N）
-                with Data.file_write_lock:
+                with data.file_write_lock:
                     with open(yon_log_file_path, 'a', encoding='utf-8') as yon_log_file:
                         yon_log_file.write('{\n')
                         yon_log_file.write(f'    "title": "{title}",\n')
@@ -59,7 +59,7 @@ def process_paper_batch(paper_indices, rq, keywords, requirements, api_key, thre
                     relevant_count += 1
                     
                     # 使用线程锁保护文件写入
-                    with Data.file_write_lock:
+                    with data.file_write_lock:
                         # 将entry写入结果文件（追加模式）
                         with open(result_file_path, 'a', encoding='utf-8') as result_file:
                             result_file.write(entry + "\n\n")
@@ -99,11 +99,11 @@ def process_papers(rq, keywords, requirements, n):
     
     # 生成带时间戳的结果文件名
     timestamp = datetime.now().strftime("%Y%m%d_%H%M")
-    Data.result_file_name = f"Result_{timestamp}"
-    result_file_path = os.path.join(result_folder, f"{Data.result_file_name}.bib")
+    data.result_file_name = f"Result_{timestamp}"
+    result_file_path = os.path.join(result_folder, f"{data.result_file_name}.bib")
     
     # 创建日志文件路径（移到Log文件夹）
-    log_file_path = os.path.join(log_folder, f"Log_{Data.result_file_name}.txt")
+    log_file_path = os.path.join(log_folder, f"Log_{data.result_file_name}.txt")
     
     # 创建Y/N判断结果日志文件路径
     yon_log_file_path = os.path.join(log_folder, f"Log_YoN_{timestamp}.txt")
@@ -111,7 +111,7 @@ def process_papers(rq, keywords, requirements, n):
     # 如果启用完整日志，创建完整日志文件
     if config.save_full_log:
         full_log_file_path = os.path.join(log_folder, f"Log_ALL_{timestamp}.txt")
-        Data.full_log_file = open(full_log_file_path, 'w', encoding='utf-8')
+        data.full_log_file = open(full_log_file_path, 'w', encoding='utf-8')
         utils.print_and_log(f"完整日志文件已创建: {full_log_file_path}")
     
     # 在结果文件开头写入搜索主题
@@ -138,13 +138,13 @@ def process_papers(rq, keywords, requirements, n):
     research_direction = rq
     
     # 处理前N篇文章（或者所有文章）
-    paper_count = len(Data.paper_data)
+    paper_count = len(data.paper_data)
     if n == -1:
         max_papers = paper_count
     else:
         max_papers = min(n, paper_count)
     
-    Data.total_papers_to_process = max_papers
+    data.total_papers_to_process = max_papers
     
     utils.print_and_log(f"\n开始处理{max_papers}篇文章的相关性...")
     utils.print_and_log(f"文章总数: {paper_count}")
@@ -171,7 +171,7 @@ def process_papers(rq, keywords, requirements, n):
     utils.print_and_log(f"实际使用线程数: {num_threads}, 平均每线程处理: {avg_papers_per_thread:.1f}篇")
     
     # 启动进度监控线程
-    Data.progress_stop_event.clear()
+    data.progress_stop_event.clear()
     progress_thread = threading.Thread(target=utils.progress_monitor, daemon=True)
     progress_thread.start()
     
@@ -179,7 +179,7 @@ def process_papers(rq, keywords, requirements, n):
     total_relevant_count = 0
     
     # 设置活跃线程数
-    Data.active_threads = num_threads  # 记录实际使用的线程数
+    data.active_threads = num_threads  # 记录实际使用的线程数
     
     with ThreadPoolExecutor(max_workers=num_threads) as executor:  # 使用动态计算的线程数
         # 提交所有任务
@@ -208,17 +208,17 @@ def process_papers(rq, keywords, requirements, n):
                 relevant_count, batch_tokens = future.result()
                 total_relevant_count += relevant_count
             except Exception as e:
-                with Data.file_write_lock:
-                    if config.save_full_log and Data.full_log_file:
-                        Data.full_log_file.write(f"\n线程{thread_id}发生错误: {str(e)}\n")
-                        Data.full_log_file.flush()
+                with data.file_write_lock:
+                    if config.save_full_log and data.full_log_file:
+                        data.full_log_file.write(f"\n线程{thread_id}发生错误: {str(e)}\n")
+                        data.full_log_file.flush()
     
     # 停止进度监控线程
-    Data.progress_stop_event.set()
+    data.progress_stop_event.set()
     progress_thread.join(timeout=2)
     
     # 计算总耗时
-    total_elapsed_time = time.time() - Data.start_time
+    total_elapsed_time = time.time() - data.start_time
     
     # 格式化总耗时
     hours = int(total_elapsed_time // 3600)
@@ -232,12 +232,12 @@ def process_papers(rq, keywords, requirements, n):
     average_time_per_paper = 1 / actual_papers_per_second if actual_papers_per_second > 0 else 0
     
     # 计算最终价格
-    from . import Price
-    final_price = Price.calculate_token_price(
-        Data.prompt_tokens_used,
-        Data.completion_tokens_used,
-        Data.prompt_cache_hit_tokens_used,
-        Data.prompt_cache_miss_tokens_used
+    from . import price
+    final_price = price.calculate_token_price(
+        data.prompt_tokens_used,
+        data.completion_tokens_used,
+        data.prompt_cache_hit_tokens_used,
+        data.prompt_cache_miss_tokens_used
     )
     
     utils.print_and_log(f"\n处理完成! 共找到{total_relevant_count}篇相关文章")
@@ -246,16 +246,16 @@ def process_papers(rq, keywords, requirements, n):
     utils.print_and_log(f"  平均单篇文章耗时: {average_time_per_paper:.2f}秒（{num_threads}个线程并发处理）")
     utils.print_and_log(f"  实际处理速度: {actual_papers_per_second:.2f}篇/秒")
     utils.print_and_log(f"----- Token使用统计 -----")
-    utils.print_and_log(f"  输入Token: {Data.prompt_tokens_used} (缓存命中: {Data.prompt_cache_hit_tokens_used}, 未命中: {Data.prompt_cache_miss_tokens_used})")
-    utils.print_and_log(f"  输出Token: {Data.completion_tokens_used}")
-    utils.print_and_log(f"  总计Token: {Data.token_used}")
+    utils.print_and_log(f"  输入Token: {data.prompt_tokens_used} (缓存命中: {data.prompt_cache_hit_tokens_used}, 未命中: {data.prompt_cache_miss_tokens_used})")
+    utils.print_and_log(f"  输出Token: {data.completion_tokens_used}")
+    utils.print_and_log(f"  总计Token: {data.token_used}")
     utils.print_and_log(f"----- 价格统计 -----")
-    utils.print_and_log(f"  总费用估算: {Price.format_price(final_price)}")
+    utils.print_and_log(f"  总费用估算: {price.format_price(final_price)}")
     utils.print_and_log(f"----- 结果文件 -----")
     utils.print_and_log(f"  相关文章已保存到: {result_file_path}")
     utils.print_and_log(f"  相关文章日志已保存到: {log_file_path}")
     utils.print_and_log(f"  所有判断结果已保存到: {yon_log_file_path}")
     
     # 关闭完整日志文件
-    if Data.full_log_file:
-        Data.full_log_file.close()
+    if data.full_log_file:
+        data.full_log_file.close()
